+++ packages/iov-bns/src/bnsconnection.ts
@@ -68,7 +68,7 @@ import {
   hashIdentifier,
   identityToAddress,
   isConfirmedWithSwapClaimOrTimeoutTransaction,
-  isConfirmedWithSwapCounterTransaction,
+  isConfirmedWithSwapOfferTransaction,
 } from "./util";
 
 const { toAscii, toHex, toUtf8 } = Encoding;
@@ -391,7 +391,7 @@ export class BnsConnection implements BcpAtomicSwapConnection {
 
     // tslint:disable-next-line:readonly-array
     const offers: OpenSwap[] = setTxs
-      .filter(isConfirmedWithSwapCounterTransaction)
+      .filter(isConfirmedWithSwapOfferTransaction)
       .map(tx => this.context.swapOfferFromTx(tx));
 
     // setTxs (esp on secondary index) may be a claim/timeout, delTxs must be a claim/timeout
@@ -420,7 +420,7 @@ export class BnsConnection implements BcpAtomicSwapConnection {
     const delTxs = this.liveTx({ tags: [bnsSwapQueryTag(query, false)] }).filter(isConfirmedTransaction);
 
     const offers: Stream<OpenSwap> = setTxs
-      .filter(isConfirmedWithSwapCounterTransaction)
+      .filter(isConfirmedWithSwapOfferTransaction)
       .map(tx => this.context.swapOfferFromTx(tx));
 
     // setTxs (esp on secondary index) may be a claim/timeout, delTxs must be a claim/timeout