+++ packages/iov-bns/src/encode.ts
@@ -7,7 +7,6 @@ import {
   SignatureBytes,
   SignedTransaction,
   SwapClaimTransaction,
-  SwapCounterTransaction,
   SwapOfferTransaction,
   SwapTimeoutTransaction,
   UnsignedTransaction,
@@ -120,8 +119,6 @@ export function buildMsg(tx: UnsignedTransaction): codecImpl.app.ITx {
       return buildSendTransaction(tx);
     case "bcp/swap_offer":
       return buildSwapOfferTx(tx);
-    case "bcp/swap_counter":
-      return buildSwapCounterTx(tx);
     case "bcp/swap_claim":
       return buildSwapClaimTx(tx);
     case "bcp/swap_timeout":
@@ -162,19 +159,6 @@ function buildSendTransaction(tx: SendTransaction): codecImpl.app.ITx {
 }
 
 function buildSwapOfferTx(tx: SwapOfferTransaction): codecImpl.app.ITx {
-  return {
-    createEscrowMsg: codecImpl.escrow.CreateEscrowMsg.create({
-      src: decodeBnsAddress(identityToAddress(tx.creator)).data,
-      arbiter: hashIdentifier(tx.hash),
-      recipient: decodeBnsAddress(tx.recipient).data,
-      amount: tx.amounts.map(encodeAmount),
-      timeout: tx.timeout,
-      memo: tx.memo,
-    }),
-  };
-}
-
-function buildSwapCounterTx(tx: SwapCounterTransaction): codecImpl.app.ITx {
   return {
     createEscrowMsg: codecImpl.escrow.CreateEscrowMsg.create({
       src: decodeBnsAddress(identityToAddress(tx.creator)).data,